fix default value move input to general panel
