+++ packages/form-js-editor/src/render/components/properties-panel/entries/DefaultValueEntry.js
@@ -1,0 +1,17 @@
+import { TextInputEntry } from '../components';
+
+
+export default function RequireCheckBox(props) {
+  const { field, editField } = props;
+
+  return (
+    <TextInputEntry
+      editField={ editField }
+      field={ field }
+      id="defaultValue"
+      label="DefaultValue"
+      path={ [ 'defaultValue' ] }
+      value={ field.key }
+    />
+  );
+}
