+++ packages/form-js-editor/src/render/components/properties-panel/entries/index.js
@@ -2,6 +2,7 @@ export { default as ActionEntry } from './ActionEntry';
 export { default as ColumnsEntry } from './ColumnsEntry';
 export { default as DescriptionEntry } from './DescriptionEntry';
 export { default as DisabledEntry } from './DisabledEntry';
+export { default as DefaultValueEntry } from './DefaultValueEntry';
 export { default as IdEntry } from './IdEntry';
 export { default as KeyEntry } from './KeyEntry';
 export { default as LabelEntry } from './LabelEntry';
